+++ services/alertsmanagement/mgmt/2018-05-05/alertsmanagement/CHANGELOG.md
@@ -1,2 +1,14 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. ActionStatus
+
+#### New Struct Fields
+
+1. Essentials.ActionStatus
+1. Essentials.Description