+++ services/preview/alertsmanagement/mgmt/2018-05-05-preview/alertsmanagement/CHANGELOG.md
@@ -1,2 +1,15 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. ActionStatus
+
+#### New Struct Fields
+
+1. AlertProperties.ActionStatus
+1. AlertProperties.AlertRule
+1. AlertProperties.Description